Total File Rename, Rearrange, and Reformat
Reformatted, rearranged,  and  renamed most files to make them easier to understand and correlate to what their functionality/purposes are.
