+++ Calculator/Search Page/alchemy_search.js
@@ -37,7 +37,7 @@ function findRecipe(page){
     xhr = new ActiveXObject("Microsoft.XMLHTTP");
   }
   let data = "recipe_name="+recipe+"&page="+pageNumber;
-  xhr.open("POST", "../../inc/Search/cook_search.php", true);
+  xhr.open("POST", "Search Tables/alchemy_search_table.php", true);
   xhr.setRequestHeader("Content-Type", "application/x-www-form-urlencoded");
   xhr.send(data);
