+++ Calculator/Search Page/cooking_recipes.php
@@ -1,31 +1,15 @@
 <!DOCTYPE html>
 
-<?php 
-session_start();
-require_once('../inc/header.php');
-$craftHeader = "";
-$craftTitle = "";
-if($_SESSION['craftType']=="Cooking"){
-    $craftTitle = "Cooking";
-    $craftHeader = "cook";
-}else if($_SESSION['craftType']=="Processing"){
-  $craftTitle = "Processing";
-  $craftHeader = "process";
-}else if($_SESSION['craftType']=="Alchemy"){
-  $craftTitle = "Alchemy";
-$craftHeader = "mix";
-}
-?>
+<?php require_once('../../Includes/header.php');?>
 
 
 <body>
-  <?php require_once('../inc/navigation.php');?>
-
+  <?php require_once('../../Includes/navigation.php');?>
 
 
   <div id="calc_jum" class="jumbotron ">
-    <h1 class="display-4 text-center"><?php  echo $craftTitle ?> Calculator</h1>
-    <p class="h5 text-center">What are you looking to <?php  echo $craftHeader ?>?</p>
+    <h1 class="display-4 text-center">Cook Calculator</h1>
+    <p class="h5 text-center">What are you looking to cook?</p>
     <div id="calc_search_bar" class="col-md-8">
       <form class="form-inline">
         <div class="input-group mb-3">
@@ -57,6 +41,6 @@ $craftHeader = "mix";
   <script  src="https://code.jquery.com/jquery-3.4.1.js"  integrity="sha256-WpOohJOqMqqyKL9FccASB9O0KwACQJpFTUBLTYOVvVU=" crossorigin="anonymous"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-  <script src="../js/Calc/search_page.js"></script>
+  <script src="cooking_search.js"></script>
 </body>
 </html>
