+++ Includes/connectToDatabase.php
