+++ Includes/search.php
@@ -11,11 +11,11 @@ if(isset($_POST['q'])){
   if(mysqli_num_rows($nav_result) > 0){
   while ($nav_row = mysqli_fetch_assoc($nav_result)) {$nav_row['table_name'];
     if($nav_row['table_name']=='cooking_recipes_table'){
-    $output .= '<li class="list_item"><span class="dropdown_list_overlay"><img src="' .$nav_row['recipe_image']. '" height="30" >' . "<a href='/Calculator/cookingDetails.php?id={$nav_row['recipe_id']}'>" .$nav_row["recipe_name"]. "</a></span></li>";
+    $output .= '<li class="list_item"><span class="dropdown_list_overlay"><img src="../' .$nav_row['recipe_image']. '" height="30" >' . "<a href='/Calculator/Recipe Page/cooking_recipe.php?id={$nav_row['recipe_id']}'>" .$nav_row["recipe_name"]. "</a></span></li>";
   }else if($nav_row['table_name']=='alchemy_recipes_table'){
-    $output .= '<li class="list_item"><span class="dropdown_list_overlay"><img src="' .$nav_row['recipe_image']. '" height="30" >' . "<a href='/Calculator/alchemyDetails.php?id={$nav_row['recipe_id']}'>" .$nav_row["recipe_name"]. "</a></span></li>";
+    $output .= '<li class="list_item"><span class="dropdown_list_overlay"><img src="../' .$nav_row['recipe_image']. '" height="30" >' . "<a href='/Calculator/Recipe Page/alchemy_recipe.php?id={$nav_row['recipe_id']}'>" .$nav_row["recipe_name"]. "</a></span></li>";
   }else if($nav_row['table_name']=='processing_recipes_table'){
-    $output .= '<li class="list_item"><span class="dropdown_list_overlay"><img src="' .$nav_row['recipe_image']. '" height="30" >' . "<a href='/Calculator/processingDetails.php?id={$nav_row['recipe_id']}'>" .$nav_row["recipe_name"]. "</a></span></li>";
+    $output .= '<li class="list_item"><span class="dropdown_list_overlay"><img src="../' .$nav_row['recipe_image']. '" height="30" >' . "<a href='/Calculator/Recipe Page/processing_recipe.php?id={$nav_row['recipe_id']}'>" .$nav_row["recipe_name"]. "</a></span></li>";
   }
