+++ index.php
@@ -1,15 +1,10 @@
 <!DOCTYPE html>
 
-<?php 
-
-session_start();
-$_SESSION['craftType'] = $_POST['sendCellValue'];
-require_once('inc/header.php');?>
+<?php require_once('Includes/header.php');?>
 
 
   <body>
-    <?php require_once('inc/navigation.php');
-    ?>
+    <?php require_once('Includes/navigation.php');?>
 
 <!-- Jumbotron -->
 <div id="index_banner" class="container-fluid">
@@ -29,7 +24,7 @@ require_once('inc/header.php');?>
 
 
     <div id="cooking_col" class="col-sm-4">
-      <div id="cooking_card" class="card text-center calculator_links">
+      <div id="cooking_card" class="card text-center">
         <div class="card-body d-flex">
           <div class="overlay">
             <h2 class="card-title mt-auto">Cooking</h2>
@@ -39,7 +34,7 @@ require_once('inc/header.php');?>
     </div>
 
     <div id="processing_col" class="col-sm-4">
-      <div id="processing_card" class="card text-center calculator_links">
+      <div id="processing_card" class="card text-center">
         <div class="card-body d-flex">
           <div class="overlay">
             <h2 class="card-title mt-auto">Processing</h2>
@@ -49,7 +44,7 @@ require_once('inc/header.php');?>
     </div>
 
     <div id="alchemy_col" class="col-sm-4">
-      <div id="alchemy_card" class="card text-center calculator_links">
+      <div id="alchemy_card" class="card text-center">
         <div class="card-body d-flex">
           <div class="overlay">
             <h2 class="card-title mt-auto">Alchemy</h2>
@@ -74,5 +69,6 @@ require_once('inc/header.php');?>
 <script  src="https://code.jquery.com/jquery-3.4.1.js"  integrity="sha256-WpOohJOqMqqyKL9FccASB9O0KwACQJpFTUBLTYOVvVU=" crossorigin="anonymous"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+<script src="/js/index_app.js"></script>
   </body>
 </html>
